address review pt2
